+++ services/confluent/mgmt/2020-03-01/confluent/CHANGELOG.md
@@ -1,2 +1,9 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. OfferDetail.MarshalJSON() ([]byte, error)
+1. OperationListResult.MarshalJSON() ([]byte, error)
+1. OrganizationResourcePropertiesOfferDetail.MarshalJSON() ([]byte, error)