+++ services/preview/confluent/mgmt/2021-03-01-preview/confluent/CHANGELOG.md
@@ -1,2 +1,7 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. OfferDetail.MarshalJSON() ([]byte, error)